@@ -23,63 +23,67 @@
   "bugs": "https://github.com/itgalaxy/eslint-plugin-itgalaxy/issues",
   "main": "index.js",
   "engines": {
-    "node": ">= 14.15.0"
+    "node": ">= 18.12.0"
   },
   "dependencies": {
-    "@babel/core": "^7.16.0",
-    "@babel/eslint-parser": "^7.16.3",
+    "@babel/core": "^7.25.2",
+    "@babel/eslint-parser": "^7.25.1",
     "confusing-browser-globals": "^1.0.10"
   },
   "devDependencies": {
     "@ava/babel": "^2.0.0",
-    "@types/confusing-browser-globals": "^1.0.0",
-    "@types/eslint": "^8.2.0",
-    "ava": "^3.15.0",
-    "eslint": "^8.6.0",
-    "eslint-config-prettier": "^8.3.0",
+    "@stylistic/eslint-plugin": "^2.6.1",
+    "@types/confusing-browser-globals": "^1.0.3",
+    "@types/eslint": "^8.56.11",
+    "ava": "^6.1.3",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
     "eslint-find-rules": "^4.0.0",
-    "eslint-plugin-ava": "^13.0.0",
-    "eslint-plugin-html": "^7.1.0",
-    "eslint-plugin-import": "^2.24.2",
-    "eslint-plugin-jest": "^27.1.4",
-    "eslint-plugin-jsdoc": "^39.6.2",
-    "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-markdown": "^3.0.0",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-react": "^7.28.0",
-    "eslint-plugin-unicorn": "^44.0.2",
-    "husky": "^8.0.1",
-    "jest": "^29.3.0",
-    "lint-staged": "^13.0.3",
+    "eslint-plugin-ava": "^14.0.0",
+    "eslint-plugin-html": "^8.1.1",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-jest": "^28.7.0",
+    "eslint-plugin-jsdoc": "^48.11.0",
+    "eslint-plugin-jsx-a11y": "^6.9.0",
+    "eslint-plugin-markdown": "^5.1.0",
+    "eslint-plugin-n": "^17.10.2",
+    "eslint-plugin-react": "^7.35.0",
+    "eslint-plugin-react-hooks": "^4.6.2",
+    "eslint-plugin-unicorn": "^55.0.0",
+    "husky": "^9.1.4",
+    "jest": "^29.7.0",
+    "lint-staged": "^15.2.7",
     "npm-run-all": "^4.1.5",
-    "nyc": "^15.1.0",
-    "prettier": "^2.4.1",
+    "nyc": "^17.0.0",
+    "prettier": "^3.3.3",
     "prop-types": "^15.7.2",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "remark-cli": "^9.0.0",
     "remark-preset-lint-itgalaxy": "^16.0.0",
     "standard-version": "^9.3.1",
-    "typescript": "^4.4.3"
+    "typescript": "^5.5.4"
   },
   "peerDependencies": {
-    "eslint": "^8.6.0",
-    "eslint-plugin-ava": "^13.0.0",
-    "eslint-plugin-html": "^7.1.0",
-    "eslint-plugin-import": "^2.24.2",
-    "eslint-plugin-jest": "^27.1.4",
-    "eslint-plugin-jsdoc": "^39.6.2",
-    "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-markdown": "^3.0.0",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-react": "^7.28.0",
-    "eslint-plugin-unicorn": "^44.0.2"
+    "@stylistic/eslint-plugin": "^2.6.1",
+    "eslint": "^8.57.0",
+    "eslint-plugin-ava": "^14.0.0",
+    "eslint-plugin-html": "^8.1.1",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-jest": "^28.7.0",
+    "eslint-plugin-jsdoc": "^48.11.0",
+    "eslint-plugin-jsx-a11y": "^6.9.0",
+    "eslint-plugin-markdown": "^5.1.0",
+    "eslint-plugin-n": "^17.10.2",
+    "eslint-plugin-react": "^7.35.0",
+    "eslint-plugin-react-hooks": "^4.6.2",
+    "eslint-plugin-unicorn": "^55.0.0"
   },
   "scripts": {
     "security": "npm audit --omit=dev",
-    "lint:check-rules": "eslint-find-rules --unused lib/config/all.js",
-    "lint:eslint-config-prettier-check": "eslint-config-prettier index.js | true",
-    "lint:prettier": "prettier . --list-different --ignore-path .gitignore",
+    "check-rules": "eslint-find-rules --unused lib/config/all.js",
+    "lint:eslint-config-prettier-check": "eslint-config-prettier index.js",
+    "lint:prettier": "prettier . --cache --list-different --ignore-path .gitignore",
     "lint:js": "eslint . --cache --ignore-path .gitignore",
     "lint:md": "remark . -f -q --ignore-path .gitignore",
     "lint:types": "tsc --pretty --noEmit",
