@@ -8,7 +8,7 @@ module.exports = {
     "jest",
     "jsx-a11y",
     "markdown",
-    "node",
+    "n",
     "react",
     "unicorn",
     "jsdoc",
