@@ -1,26 +1,24 @@
 "use strict";
 
 const possibleErrorsRules = require("./rules/possible-errors");
-const bestPracticesRules = require("./rules/best-practices");
-const strictModeRules = require("./rules/strict-mode");
-const variablesRules = require("./rules/variables");
+const suggestions = require("./rules/suggestions");
 const stylisticIssuesRules = require("./rules/stylistic-issues");
-const ecmascript6Rules = require("./rules/ecmascript-6");
+const deprecated = require("./rules/deprecated");
 const unicornRules = require("./rules/unicorn");
+const layoutAndFormatting = require("./rules/layout-and-formatting");
 
 module.exports = {
   parserOptions: {
     ecmaVersion: 2022,
   },
   env: { es2020: true },
-  plugins: ["unicorn"],
+  plugins: ["unicorn", "@stylistic"],
   rules: {
     ...possibleErrorsRules,
-    ...bestPracticesRules,
-    ...strictModeRules,
-    ...variablesRules,
+    ...suggestions,
+    ...layoutAndFormatting,
+    ...deprecated,
     ...stylisticIssuesRules,
-    ...ecmascript6Rules,
     ...unicornRules,
   },
 };
