@@ -41,8 +41,8 @@ module.exports = {
         "no-unused-expressions": "off",
         "import/no-unresolved": "off",
         "import/no-extraneous-dependencies": "off",
-        "node/no-unpublished-require": "off",
-        "node/no-unpublished-import": "off",
+        "n/no-unpublished-require": "off",
+        "n/no-unpublished-import": "off",
 
         // Adding a "use strict" directive at the top of every code block is tedious and distracting. The config opts into strict mode parsing without the directive.
         strict: "off",
