@@ -50,7 +50,7 @@ module.exports = {
     setInterval: "readonly",
     setTimeout: "readonly",
   },
-  plugins: ["node", "unicorn"],
+  plugins: ["n", "unicorn"],
   settings: {
     node: {
       tryExtensions: [".mjs", ".cjs", ".js", ".jsx", ".json", ".node"],
