@@ -44,8 +44,7 @@ module.exports = {
   // Allow only use of the asserts that have no power-assert alternative.
   "ava/prefer-power-assert": "off",
   // Prefer using `t.regex()` to test regular expressions
-  // TODO Enable in future
-  "ava/prefer-t-regex": "off",
+  "ava/prefer-t-regex": "error",
   // Ensure tests have a title.
   "ava/test-title": "error",
   // Ensure test titles have a certain format
