@@ -4,19 +4,21 @@
 
 module.exports = {
   // Enforce emojis are wrapped in and provide screenreader access.
-  "jsx-a11y/accessible-emoji": "error",
+  // Deprecated
+  "jsx-a11y/accessible-emoji": "off",
   // Enforce all elements that require alternative text have meaningful information to relay back to end user.
   "jsx-a11y/alt-text": [
     "error",
     {
       elements: ["img", "object", "area", 'input[type="image"]'],
-      img: [],
-      object: [],
-      area: [],
-      'input[type="image"]': [],
+      img: ["Image"],
+      object: ["Object"],
+      area: ["Area"],
+      'input[type="image"]': ["InputImage"],
     },
   ],
-
+  // Enforce <a> text to not exactly match "click here", "here", "link", or "a link".
+  "jsx-a11y/anchor-ambiguous-text": "off",
   // Enforce all anchors to contain accessible content.
   "jsx-a11y/anchor-has-content": ["error", { components: [] }],
   // Enforce all anchors are valid, navigable elements.
@@ -73,7 +75,6 @@ module.exports = {
     },
   ],
   // Enforce that a label tag has a text label and an associated control.
-  // TODO Maybe enable in future
   "jsx-a11y/control-has-associated-label": "off",
   // Enforce that a label tag has a text label and an associated control.
   "jsx-a11y/label-has-associated-control": [
@@ -101,6 +102,8 @@ module.exports = {
   "jsx-a11y/mouse-events-have-key-events": "error",
   // Enforce that the accessKey prop is not used on any element to avoid complications with keyboard commands used by a screenreader.
   "jsx-a11y/no-access-key": "error",
+  // Disallow aria-hidden="true" from being set on focusable elements.
+  "jsx-a11y/no-aria-hidden-on-focusable": "error",
   // Enforce autoFocus prop is not used.
   "jsx-a11y/no-autofocus": ["error", { ignoreNonDom: false }],
   // Enforce distracting elements are not used.
@@ -174,7 +177,8 @@ module.exports = {
     },
   ],
   // Enforce that onBlur is used instead of onChange.
-  "jsx-a11y/no-onchange": "error",
+  // Deprecated
+  "jsx-a11y/no-onchange": "off",
   // Enforce explicit role property is not the same as implicit/default role property on element.
   "jsx-a11y/no-redundant-roles": "error",
   // Enforce non-interactive elements have no interactive handlers.
@@ -191,6 +195,8 @@ module.exports = {
       ],
     },
   ],
+  // Enforces using semantic DOM elements over the ARIA role property.
+  "jsx-a11y/prefer-tag-over-role": "error",
   // Enforce that elements with ARIA roles must have all required attributes for that role.
   "jsx-a11y/role-has-required-aria-props": "error",
   // Enforce that elements with explicit or implicit roles defined contain only aria-* properties supported by that role.
