@@ -12,6 +12,8 @@ module.exports = {
   ],
   // Prevent usage of `button` elements without an explicit `type` attribute
   "react/button-has-type": "error",
+  // This rule enforces `onChange` or `readonly` attribute for checked property of input elements.
+  "react/checked-requires-onchange-or-readonly": "error",
   // Enforce all defaultProps have a corresponding non-required PropType
   "react/default-props-match-prop-types": [
     "error",
@@ -39,7 +41,6 @@ module.exports = {
   // Enforce a specific function type for function components
   "react/function-component-definition": "off",
   // This rule checks whether the value and setter variables destructured from a React.useState() call are named symmetrically.
-  // TODO add `allowDestructuredState`
   "react/hook-use-state": "error",
   // The sandbox attribute enables an extra set of restrictions for the content in the iframe. Using sandbox attribute is considered a good security practice.
   "react/iframe-missing-sandbox": "error",
@@ -80,6 +81,8 @@ module.exports = {
   ],
   // Enforce that namespaces are not used in React elements
   "react/no-namespace": "error",
+  // Warns if in a functional component, an object type value (such as array/object literal/function/etc) is used as default prop, to prevent potential unnecessary rerenders, and performance regressions.
+  "react/no-object-type-as-default-prop": "off",
   // Prevent usage of shouldComponentUpdate when extending React.PureComponent
   "react/no-redundant-should-component-update": "error",
   // Prevent usage of the return value of `React.render`.
@@ -192,6 +195,8 @@ module.exports = {
       },
     },
   ],
+  // This rule checks all components and verifies that all defaultProps declarations are sorted alphabetically. A spread attribute resets the verification. The default configuration of the rule is case-sensitive.
+  "react/sort-default-props": "off",
   // Enforce propTypes declarations alphabetical sorting
   "react/sort-prop-types": [
     "off",
@@ -315,6 +320,8 @@ module.exports = {
   // Enforces that there is exactly one space between all attributes and after tag name and the first attribute in the same line.
   // Disable in favor `prettier`
   "react/jsx-props-no-multi-spaces": "off",
+  // Enforces that any unique expression is only spread once. Generally spreading the same expression twice is an indicator of a mistake since any attribute between the spreads may be overridden when the intent was not to. Even when that is not the case this will lead to unnecessary computations being performed.
+  "react/jsx-props-no-spread-multi": "error",
   // Disallow JSX props spreading
   "react/jsx-props-no-spreading": "off",
   // Enforce defaultProps declarations alphabetical sorting
