@@ -4,298 +4,154 @@
 
 module.exports = {
   // Enforce linebreaks after opening and before closing array brackets
-  // Prettier does the job
-  "array-bracket-newline": "off",
+  "@stylistic/array-bracket-newline": "off",
   // Enforce consistent spacing inside array brackets
-  // Prettier does the job
-  "array-bracket-spacing": "off",
+  "@stylistic/array-bracket-spacing": "off",
   // Enforce line breaks after each array element
-  // Prettier does the job
-  "array-element-newline": "off",
+  "@stylistic/array-element-newline": "off",
+  // Require parentheses around arrow function arguments
+  "@stylistic/arrow-parens": "off",
+  // Enforce consistent spacing before and after the arrow in arrow functions
+  "@stylistic/arrow-spacing": "off",
   // Enforce line breaks after each array element
-  // Prettier does the job
-  "block-spacing": "off",
+  "@stylistic/block-spacing": "off",
   // Enforce consistent spacing inside single-line blocks
-  // Prettier does the job
-  "brace-style": "off",
-  // Require camel case names
-  camelcase: [
-    "error",
-    {
-      properties: "always",
-    },
-  ],
-  // Enforce or disallow capitalization of the first letter of a comment
-  // wHO CaRes?
-  "capitalized-comments": "off",
+  "@stylistic/brace-style": "off",
   // Require or disallow trailing commas
-  // Prettier does the job
-  "comma-dangle": "off",
+  "@stylistic/comma-dangle": "off",
   // Enforce spacing before and after comma
-  // Prettier does the job
-  "comma-spacing": "off",
+  "@stylistic/comma-spacing": "off",
   // Enforce one true comma style
-  // Prettier does the job
-  "comma-style": "off",
+  "@stylistic/comma-style": "off",
   // Require or disallow padding inside computed properties
-  // Prettier does the job
-  "computed-property-spacing": "off",
-  // Enforces consistent naming when capturing the current execution context
-  "consistent-this": ["error", "self"],
+  "@stylistic/computed-property-spacing": "off",
+  // Enforce consistent newlines before and after dots
+  "@stylistic/dot-location": "off",
   // Enforce newline at the end of file, with no multiple empty lines
-  // Prettier does the job
-  "eol-last": "off",
+  "@stylistic/eol-last": "off",
   // Require or disallow spacing between function identifiers and their invocations
-  // Prettier does the job
-  "func-call-spacing": "off",
-  // Requires function names to match the name of the variable or property to which they are
-  "func-name-matching": [
-    "error",
-    "always",
-    {
-      considerPropertyDescriptor: true,
-      includeCommonJSModuleExports: false,
-    },
-  ],
-  // Don"t require function expressions to have a name
-  // TODO maybe enable in future
-  "func-names": "off",
-  // Enforces use of function declarations or expressions
-  "func-style": [
-    "error",
-    "declaration",
-    {
-      allowArrowFunctions: true,
-    },
-  ],
+  "@stylistic/func-call-spacing": "off",
   // Enforce line breaks between arguments of a function call
-  // Prettier does the job
-  "function-call-argument-newline": "off",
+  "@stylistic/function-call-argument-newline": "off",
+  // Require or disallow spacing between function identifiers and their invocations
+  "@stylistic/function-call-spacing": "off",
   // Enforce consistent line breaks inside function parentheses
-  // Prettier does the job
-  "function-paren-newline": "off",
-  // Blacklist certain identifiers to prevent them being used
-  // No need
-  "id-denylist": "off",
-  // This option enforces minimum and maximum identifier lengths (variable names, property names etc.)
-  "id-length": [
-    "error",
-    {
-      min: 2,
-      max: Number.POSITIVE_INFINITY,
-      properties: "always",
-      exceptions: [
-        // jQuery
-        "$",
-        // Loops
-        "i",
-        "j",
-        "k",
-        "v",
-        "m",
-        "n",
-        "t",
-        // Lodash
-        "_",
-        // Comparison
-        "a",
-        "b",
-      ],
-    },
-  ],
-  // Require identifiers to match the provided regular expression
-  "id-match": [
-    "error",
-    "^[$a-zA-Z_][a-zA-Z0-9_]*$",
-    {
-      properties: true,
-    },
-  ],
+  "@stylistic/function-paren-newline": "off",
+  // Enforce consistent spacing around `*` operators in generator functions
+  "@stylistic/generator-star-spacing": "off",
   // Enforce the location of arrow function bodies
-  // Prettier does the job
-  "implicit-arrow-linebreak": "off",
+  "@stylistic/implicit-arrow-linebreak": "off",
   // This option sets a specific tab width for your code
-  // Prettier does the job
-  indent: "off",
-  // Specify whether double or single quotes should be used in JSX attributes
-  // Prettier does the job
-  "jsx-quotes": "off",
+  "@stylistic/indent": "off",
+  // Indentation for binary operators
+  "@stylistic/indent-binary-ops": "off",
+  // Enforce or disallow spaces inside of curly braces in JSX attributes and expressions
+  "@stylistic/jsx-child-element-spacing": "off",
+  // Enforce closing bracket location in JSX
+  "@stylistic/jsx-closing-bracket-location": "off",
+  // Enforce closing tag location for multiline JSX
+  "@stylistic/jsx-closing-tag-location": "off",
+  // Disallow unnecessary JSX expressions when literals alone are sufficient or enforce JSX expressions on literals in JSX children or attributes
+  "@stylistic/jsx-curly-brace-presence": "off",
+  // Enforce consistent linebreaks in curly braces in JSX attributes and expressions
+  "@stylistic/jsx-curly-newline": "off",
+  // Enforce or disallow spaces inside of curly braces in JSX attributes and expressions
+  "@stylistic/jsx-curly-spacing": "off",
+  // Enforce or disallow spaces around equal signs in JSX attributes
+  "@stylistic/jsx-equals-spacing": "off",
+  // Enforce proper position of the first property in JSX
+  "@stylistic/jsx-first-prop-new-line": "off",
+  // Enforce line breaks before and after JSX elements when they are used as arguments to a function.
+  "@stylistic/jsx-function-call-newline": "off",
+  // Enforce JSX indentation. Deprecated, use `indent` rule instead.
+  "@stylistic/jsx-indent": "off",
+  // Enforce props indentation in JSX
+  "@stylistic/jsx-indent-props": "off",
+  // Enforce maximum of props on a single line in JSX
+  "@stylistic/jsx-max-props-per-line": "off",
+  // Require or prevent a new line after jsx elements and expressions.
+  "@stylistic/jsx-newline": "off",
+  // Require one JSX element per line
+  "@stylistic/jsx-one-expression-per-line": "off",
+  // Enforce PascalCase for user-defined JSX components
+  "@stylistic/jsx-pascal-case": "off",
+  // Disallow multiple spaces between inline JSX props
+  "@stylistic/jsx-props-no-multi-spaces": "off",
+  // Enforce the consistent use of either double or single quotes in JSX attributes
+  "@stylistic/jsx-quotes": "off",
+  // Disallow extra closing tags for components without children
+  "@stylistic/jsx-self-closing-comp": "off",
+  // Enforce props alphabetical sorting
+  "@stylistic/jsx-sort-props": "off",
+  // Enforce whitespace in and around the JSX opening and closing brackets
+  "@stylistic/jsx-tag-spacing": "off",
+  // Disallow missing parentheses around multiline JSX
+  "@stylistic/jsx-wrap-multilines": "off",
   // Enforces spacing between keys and values in object literal properties
-  // Prettier does the job
-  "key-spacing": "off",
+  "@stylistic/key-spacing": "off",
   // Enforce spacing before and after keywords
-  // Prettier does the job
-  "keyword-spacing": "off",
+  "@stylistic/keyword-spacing": "off",
   // Enforce position of line comments
-  // Prettier does the job
-  "line-comment-position": "off",
+  "@stylistic/line-comment-position": "off",
   // Disallow mixed "LF" and "CRLF" as linebreaks
-  // Prettier does the job
-  "linebreak-style": "off",
+  "@stylistic/linebreak-style": "off",
   // Enforces empty lines around comments
-  // Prettier does the job
-  "lines-around-comment": "off",
+  "@stylistic/lines-around-comment": "off",
   // Require or disallow an empty line between class members
-  // Prettier does the job
-  "lines-between-class-members": "error",
-  // Enforce a maximum depth that blocks can be nested
-  // No need
-  "max-depth": "off",
+  "@stylistic/lines-between-class-members": "error",
   // Enforce a maximum line length
-  // No need
-  "max-len": "off",
-  // Enforce a maximum number of lines per file
-  // No need
-  "max-lines": "off",
-  // Enforce a maximum number of line of code in a function
-  // No need
-  "max-lines-per-function": "off",
-  // Specify the maximum depth callbacks can be nested
-  // No need
-  "max-nested-callbacks": "off",
-  // Enforce a maximum number of parameters in function definitions
-  "max-params": ["error", 7],
+  "@stylistic/max-len": "off",
   // Enforce a maximum number of statements allowed per line
-  "max-statements-per-line": [
-    "error",
-    {
-      max: 1,
-    },
-  ],
+  "@stylistic/max-statements-per-line": "off",
+  // Require a specific member delimiter style for interfaces and type literals
+  "@stylistic/member-delimiter-style": "off",
   // Enforce a particular style for multiline comments
   // Comments can be different
-  "multiline-comment-style": "off",
-  // Enforce a maximum number of statements allowed in function blocks
-  // No need
-  "max-statements": "off",
+  "@stylistic/multiline-comment-style": "off",
   // Enforce newlines between operands of ternary expressions
-  // No need
-  "multiline-ternary": "off",
-  // Require a capital letter for constructors
-  "new-cap": [
-    "error",
-    {
-      newIsCap: true,
-      newIsCapExceptions: [],
-      capIsNew: true,
-      capIsNewExceptions: [
-        // Immutable
-        "Immutable.Map",
-        "Immutable.Set",
-        "Immutable.List",
-
-        // jQuery
-        "jQuery.Animation",
-        "$.Animation",
-        "jQuery.Callbacks",
-        "$.Callbacks",
-        "jQuery.Deferred",
-        "$.Deferred",
-        "jQuery.Event",
-        "$.Event",
-        "jQuery.Tween",
-        "$.Tween",
-      ],
-      properties: true,
-    },
-  ],
+  "@stylistic/multiline-ternary": "off",
   // Disallow the omission of parentheses when invoking a constructor with no arguments
-  // Prettier does the job
-  "new-parens": "off",
+  "@stylistic/new-parens": "off",
   // Enforce newline after each call when chaining the calls
-  // Prettier does the job
-  "newline-per-chained-call": "off",
-  // Disallow use of the Array constructor
-  // Disable in favor `unicorn/no-new-array`
-  "no-array-constructor": "off",
-  // Disallow bitwise operators
-  "no-bitwise": "error",
-  // Disallow use of the continue statement
-  // Continue can be useful in many cases
-  "no-continue": "off",
-  // Disallow comments inline after code
-  // It's not a problem
-  "no-inline-comments": "off",
-  // Disallow if as the only statement in an else block
-  "no-lonely-if": "error",
+  "@stylistic/newline-per-chained-call": "off",
+  // Disallow arrow functions where they could be confused with comparisons
+  "@stylistic/no-confusing-arrow": "off",
+  // Disallow unnecessary parentheses
+  "@stylistic/no-extra-parens": "off",
+  // Disallow unnecessary semicolons
+  "@stylistic/no-extra-semi": "off",
+  // Disallow leading or trailing decimal points in numeric literals
+  "@stylistic/no-floating-decimal": "off",
   // Disallow mixed spaces and tabs for indentation
-  // No need
-  "no-mixed-operators": "off",
+  "@stylistic/no-mixed-operators": "off",
   // Disallow mixed spaces and tabs for indentation
-  // Prettier does the job
-  "no-mixed-spaces-and-tabs": "off",
-  // Disallow use of chained assignment expressions
-  "no-multi-assign": "error",
+  "@stylistic/no-mixed-spaces-and-tabs": "off",
+  // Disallow multiple spaces
+  "@stylistic/no-multi-spaces": "off",
   // Disallow multiple empty lines
-  // Prettier does the job
-  "no-multiple-empty-lines": "off",
-  // Disallow negated conditions
-  // No need
-  "no-negated-condition": "off",
-  // Disallow nested ternary expressions
-  // Sometimes useful, but requires good tuning, but `eslint` doesn't provide it
-  "no-nested-ternary": "off",
-  // Disallow use of the Object constructor
-  "no-new-object": "error",
-  // Disallow space between function identifier and application
-  "no-plusplus": [
-    "error",
-    {
-      allowForLoopAfterthoughts: true,
-    },
-  ],
-  // Disallow specified syntax
-  "no-restricted-syntax": [
-    "error",
-    "DebuggerStatement",
-    "LabeledStatement",
-    "WithStatement",
-  ],
+  "@stylistic/no-multiple-empty-lines": "off",
   // Disallow all tabs
-  // Prettier does the job
-  "no-tabs": "off",
-  // Disallow the use of ternary operators
-  // No need
-  "no-ternary": "off",
+  "@stylistic/no-tabs": "off",
   // Disallow trailing whitespace at the end of lines
-  // Prettier does the job
-  "no-trailing-spaces": "off",
-  // Allow dangling underscores in identifiers
-  // Preferable to use `private` properties and methods
-  "no-underscore-dangle": "off",
-  // Disallow the use of Boolean literals in conditional expressions
-  // Prefer `a || b` over `a ? a : b`
-  "no-unneeded-ternary": ["error", { defaultAssignment: false }],
+  "@stylistic/no-trailing-spaces": "off",
   // Disallow whitespace before properties
-  // Prettier does the job
-  "no-whitespace-before-property": "off",
+  "@stylistic/no-whitespace-before-property": "off",
   // Disallow whitespace before properties
-  // Prettier does the job
-  "nonblock-statement-body-position": "off",
+  "@stylistic/nonblock-statement-body-position": "off",
   // Disallow dangling underscores in identifiers
-  // Prettier does the job
-  "object-curly-newline": "off",
+  "@stylistic/object-curly-newline": "off",
   // Require or disallow padding inside curly braces
-  // Prettier does the job
-  "object-curly-spacing": "off",
+  "@stylistic/object-curly-spacing": "off",
   // Enforce placing object properties on separate lines
-  // Prettier does the job
-  "object-property-newline": "off",
-  // Require or disallow an newline around variable declarations
-  // Prettier does the job
-  "one-var-declaration-per-line": "off",
-  // Allow or disallow one variable declaration per function
-  "one-var": ["error", "never"],
-  // Require assignment operator shorthand where possible or prohibit it entirely
-  "operator-assignment": ["error", "always"],
-  // Enforce operators to be placed before or after line breaks
-  // Prettier does the job
-  "operator-linebreak": "off",
+  "@stylistic/object-property-newline": "off",
+  // Require or disallow newlines around variable declarations
+  "@stylistic/one-var-declaration-per-line": "off",
+  // Enforce consistent linebreak style for operators
+  "@stylistic/operator-linebreak": "off",
   // Enforce padding within blocks
-  // Prettier does the job
-  "padded-blocks": "off",
+  "@stylistic/padded-blocks": "off",
   // Require or disallow padding lines between statements
-  "padding-line-between-statements": [
+  "@stylistic/padding-line-between-statements": [
     "error",
     // Statement `block` does't need to be covered
     // Newline between `block-like` (this matches statements that the last token is the closing brace of blocks)
@@ -377,51 +233,33 @@ module.exports = {
       next: ["import"],
     },
   ],
-  // Disallow the use of Math.pow in favor of the ** operator
-  "prefer-exponentiation-operator": "error",
-  // Disallow using Object.assign with an object literal as the first argument and prefer the use of object spread instead.
-  "prefer-object-spread": "error",
   // Require quotes around object literal property names
-  // Prettier does the job
-  "quote-props": "off",
+  "@stylistic/quote-props": "off",
   // Specify whether backticks, double or single quotes should be used
-  quotes: [
+  "@stylistic/quotes": [
     "error",
     "double",
     { avoidEscape: true, allowTemplateLiterals: false },
   ],
-  // Enforce spacing before and after semicolons
-  // Prettier does the job
-  "semi-spacing": "off",
-  // Enforce location of semicolons
-  // Prettier does the job
-  "semi-style": "off",
+  "@stylistic/rest-spread-spacing": "off",
   // Require or disallow use of semicolons instead of ASI
-  // Prettier does the job
-  semi: "off",
-  // Requires object keys to be sorted
-  // No need
-  "sort-keys": "off",
-  // Sort variables within the same declaration block
-  // No need
-  "sort-vars": "off",
+  "@stylistic/semi": "off",
+  // Enforce spacing before and after semicolons
+  "@stylistic/semi-spacing": "off",
+  // Enforce location of semicolons
+  "@stylistic/semi-style": "off",
   // Require or disallow space before blocks
-  // Prettier does the job
-  "space-before-blocks": "off",
+  "@stylistic/space-before-blocks": "off",
   // Require or disallow space before function opening parenthesis
-  // Prettier does the job
-  "space-before-function-paren": "off",
+  "@stylistic/space-before-function-paren": "off",
   // Require or disallow spaces inside parentheses
-  // Prettier does the job
-  "space-in-parens": "off",
+  "@stylistic/space-in-parens": "off",
   // Require spaces around operators
-  // Prettier does the job
-  "space-infix-ops": "off",
+  "@stylistic/space-infix-ops": "off",
   // Require or disallow spaces before/after unary operators (words on by default, nonwords)
-  // Prettier does the job
-  "space-unary-ops": "off",
+  "@stylistic/space-unary-ops": "off",
   // Require or disallow a space immediately following the // or /* in a comment
-  "spaced-comment": [
+  "@stylistic/spaced-comment": [
     "error",
     "always",
     {
@@ -437,15 +275,21 @@ module.exports = {
     },
   ],
   // Enforce spacing around colons of switch statements
-  // Prettier does the job
-  "switch-colon-spacing": "off",
+  "@stylistic/switch-colon-spacing": "off",
+  // Require or disallow spacing around embedded expressions of template strings
+  "@stylistic/template-curly-spacing": "off",
   // Require or disallow spacing between template tags and their literals
-  // Prettier does the job
-  "template-tag-spacing": "off",
-  // Require or disallow Unicode byte order mark (BOM)
-  // Prettier keep Unicode byte order mark
-  "unicode-bom": ["error", "never"],
+  "@stylistic/template-tag-spacing": "off",
+  // Require consistent spacing around type annotations
+  "@stylistic/type-annotation-spacing": "off",
+  // Enforces consistent spacing inside TypeScript type generics
+  "@stylistic/type-generic-spacing": "off",
+  // Expect space before the type declaration in the named tuple
+  "@stylistic/type-named-tuple-spacing": "off",
+  // Require parenthesis around regex literals
+  "@stylistic/wrap-iife": "off",
   // Require regex literals to be wrapped in parentheses
-  // Prettier does the job
-  "wrap-regex": "off",
+  "@stylistic/wrap-regex": "off",
+  // Require or disallow spacing around the `*` in `yield*` expressions
+  "@stylistic/yield-star-spacing": "off",
 };
