@@ -10,6 +10,7 @@ const scriptSharedConfig = {
     ecmaFeatures: { globalReturn: true },
   },
   globals: {
+    //  ...globals.es2021,
     __dirname: "readonly",
     __filename: "readonly",
     exports: "writable",
@@ -23,7 +24,7 @@ const scriptSharedConfig = {
       node: { extensions: [".mjs", ".cjs", ".js", ".jsx", ".json", ".node"] },
     },
   },
-  plugins: ["import", "node"],
+  plugins: ["import", "n"],
   rules: {
     ...importRules,
     "import/extensions": [
@@ -47,33 +48,26 @@ const scriptSharedConfig = {
     ],
 
     // Possible Errors
-    "node/no-exports-assign": "error",
-    "node/no-extraneous-require": "off",
-    "node/no-missing-require": "off",
-    "node/no-new-require": "error",
-    "node/no-unpublished-require": "error",
-    "node/no-unsupported-features/es-syntax": [
+    "n/no-exports-assign": "error",
+    "n/no-extraneous-require": "off",
+    "n/no-missing-require": "off",
+    "n/no-new-require": "error",
+    "n/no-unpublished-require": "error",
+    "n/no-unsupported-features/es-syntax": [
       "error",
       { ignores: ["dynamicImport"] },
     ],
 
     // Stylistic Issues
-    "node/exports-style": [
-      "error",
-      "module.exports",
-      {
-        allowBatchAssign: false,
-      },
-    ],
-    "node/global-require": "off",
-    "node/no-mixed-requires": [
+    "n/exports-style": "error",
+    "n/no-mixed-requires": [
       "error",
       {
         grouping: true,
         allowCall: true,
       },
     ],
-    "node/no-restricted-require": "off",
+    "n/no-restricted-require": "off",
   },
 };
 
