@@ -23,44 +23,37 @@ const dirtySharedConfig = {
       node: { extensions: [".mjs", ".cjs", ".js", ".jsx", ".json", ".node"] },
     },
   },
-  plugins: ["import", "node"],
+  plugins: ["import", "n"],
   rules: {
     ...importRules,
     "import/extensions": "off",
 
     // Possible Errors
-    "node/no-exports-assign": "error",
-    "node/no-extraneous-require": "off",
-    "node/no-extraneous-import": "off",
-    "node/no-missing-require": "off",
-    "node/no-missing-import": "off",
-    "node/no-new-require": "error",
-    "node/no-unpublished-require": "error",
-    "node/no-unpublished-import": "error",
-    "node/no-unsupported-features/es-syntax": [
+    "n/no-exports-assign": "error",
+    "n/no-extraneous-require": "off",
+    "n/no-extraneous-import": "off",
+    "n/no-missing-require": "off",
+    "n/no-missing-import": "off",
+    "n/no-new-require": "error",
+    "n/no-unpublished-require": "error",
+    "n/no-unpublished-import": "error",
+    "n/no-unsupported-features/es-syntax": [
       "error",
       { ignores: ["modules", "dynamicImport"] },
     ],
 
     // Stylistic Issues
-    "node/exports-style": [
-      "error",
-      "module.exports",
-      {
-        allowBatchAssign: false,
-      },
-    ],
-    "node/file-extension-in-import": "off",
-    "node/global-require": "off",
-    "node/no-mixed-requires": [
+    "n/exports-style": "error",
+    "n/file-extension-in-import": "off",
+    "n/no-mixed-requires": [
       "error",
       {
         grouping: true,
         allowCall: true,
       },
     ],
-    "node/no-restricted-import": "off",
-    "node/no-restricted-require": "off",
+    "n/no-restricted-import": "off",
+    "n/no-restricted-require": "off",
 
     "no-import-assign": "error",
   },
