@@ -17,6 +17,7 @@ const moduleSharedConfig = {
     },
   },
   globals: {
+    // ...globals.es2021,
     __dirname: "off",
     __filename: "off",
     exports: "off",
@@ -30,7 +31,7 @@ const moduleSharedConfig = {
       node: { extensions: [".mjs", ".cjs", ".js", ".jsx", ".json", ".node"] },
     },
   },
-  plugins: ["import", "node"],
+  plugins: ["import", "n"],
   rules: {
     ...importRules,
     "import/extensions": [
@@ -46,19 +47,20 @@ const moduleSharedConfig = {
         noUselessIndex: false,
       },
     ],
+    "import/no-import-module-exports": "error",
 
     // Possible Errors
-    "node/no-extraneous-import": "off",
-    "node/no-missing-import": "off",
-    "node/no-unpublished-import": "error",
-    "node/no-unsupported-features/es-syntax": [
+    "n/no-extraneous-import": "off",
+    "n/no-missing-import": "off",
+    "n/no-unpublished-import": "error",
+    "n/no-unsupported-features/es-syntax": [
       "error",
       { ignores: ["modules", "dynamicImport"] },
     ],
 
     // Stylistic Issues
-    "node/file-extension-in-import": "off",
-    "node/no-restricted-import": "off",
+    "n/file-extension-in-import": "off",
+    "n/no-restricted-import": "off",
 
     "no-import-assign": "error",
   },
