@@ -2,6 +2,7 @@
 
 const jsxA11yRules = require("../rules/jsx-a11y");
 const reactRules = require("../rules/react");
+const reactHooksRules = require("../rules/react-hooks");
 
 const reactSharedConfig = {
   parserOptions: {
@@ -25,10 +26,11 @@ const reactSharedConfig = {
       version: "detect",
     },
   },
-  plugins: ["react", "jsx-a11y"],
+  plugins: ["react", "jsx-a11y", "react-hooks"],
   rules: {
     ...reactRules,
     ...jsxA11yRules,
+    ...reactHooksRules,
   },
 };
 
